requirements.txt
python-telegram-bot==13.5
expiringdict==1.2.1
google-api-python-client==2.12.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4